Revert "Extract SwaggerReporter to fordefi_swagger_reporter sub-package [DEV-24549]"
