--- fordefi_swagger_reporter/README.md
@@ -1,38 +1,0 @@
-# fordefi_swagger_reporter
-
-Runtime reporter abstraction for Dart clients generated by the fordefi fork of
-`swagger_dart_code_generator`.
-
-This is a tiny pure-Dart package (~50 LOC). The generator emits calls like
-`SwaggerReporterHelper.report(...)` into generated chopper clients and model
-factories; this package provides the `SwaggerReporter` interface and the static
-`SwaggerReporterHelper` dispatcher so the generated output is self-contained and
-publishable.
-
-## Usage
-
-Add as a runtime dependency wherever you consume generated swagger clients:
-
-```yaml
-dependencies:
-  fordefi_swagger_reporter: ^1.0.0
-```
-
-At app startup, register your reporter implementation:
-
-```dart
-import 'package:fordefi_swagger_reporter/fordefi_swagger_reporter.dart';
-
-class _FLogReporter implements SwaggerReporter {
-  @override
-  void report(String msg) => FLog.error(text: msg);
-}
-
-void main() {
-  SwaggerReporterHelper.setReporter(_FLogReporter());
-  // ...
-}
-```
-
-Until a reporter is registered, all `SwaggerReporterHelper.report` calls are
-no-ops.
