--- fordefi_swagger_reporter/pubspec.yaml
@@ -1,12 +1,0 @@
-name: fordefi_swagger_reporter
-description: >
-  Runtime reporter abstraction for Dart clients generated by the fordefi fork
-  of swagger_dart_code_generator. Tiny pure-Dart package — generated code
-  imports it to emit log/report calls; consumers wire it to their own logger
-  (e.g. FLog) at startup.
-version: 1.0.0
-homepage: https://github.com/arnac-io/swagger-dart-code-generator/tree/fordefi_master/fordefi_swagger_reporter
-publish_to: https://fordefi.jfrog.io/artifactory/api/pub/mobile-fordefi-packages
-
-environment:
-  sdk: ">=3.10.4 <4.0.0"
